src/pyproject.toml
[tool.poetry]
name = "two-captains-online-store"
version = "0.1.0"
description = ""
authors = ["kholkin.anthony <kholkin.anthony@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standart"], version = "^0.30.1"}
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.3.1"
loguru = "^0.7.2"
aioboto3 = "^13.0.1"
itsdangerous = "^2.2.0"
ruff = "^0.5.0"
mypy = "^1.10.1"
alembic = "^1.13.2"
inflection = "^0.5.1"
asyncpg = "^0.29.0"
fastapi-pagination = "^0.12.26"
fastapi-filter = "^2.0.0"
pillow = "^10.4.0"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120